sha1-checked: implement hardware acceleration on arm64 and x86_64
Fixes #909.
