+++ sha1-checked/Cargo.toml
@@ -18,9 +18,13 @@ exclude = [
 ]
 
 [dependencies]
+cfg-if = "1.0"
 digest = "0.11"
 sha1 = { version = "0.11.0-rc.5", default-features = false }
 zeroize = { version = "1.8", default-features = false, optional = true }
+
+[target.'cfg(any(target_arch = "aarch64", target_arch = "x86", target_arch = "x86_64"))'.dependencies]
+cpufeatures = "0.3"
 
 [dev-dependencies]
 digest = { version = "0.11", features = ["dev"] }
