+++ sha1-checked/src/compress/backend/aarch64_sha2.rs
@@ -1,0 +1,189 @@
+//! SHA-1 `aarch64` backend, spilling the message schedule alongside the
+//! digest.
+
+use crate::compress::soft::K;
+
+#[cfg(not(target_arch = "aarch64"))]
+compile_error!("aarch64_sha2 backend can be used only on aarch64 target arches");
+
+#[target_feature(enable = "sha2")]
+#[allow(unsafe_op_in_unsafe_fn)]
+#[allow(clippy::too_many_lines)]
+pub(crate) unsafe fn compress_spill(state: &mut [u32; 5], block: &[u8; 64], w: &mut [u32; 80]) {
+    use core::arch::aarch64::*;
+
+    let mut abcd = vld1q_u32(state.as_ptr());
+    let mut e0 = state[4];
+    let [k0, k1, k2, k3] = K.map(|k| vdupq_n_u32(k));
+    let (mut e1, mut tmp0, mut tmp1);
+
+    let abcd_cpy = abcd;
+    let e0_cpy = e0;
+    let wp = w.as_mut_ptr();
+
+    // Load and reverse byte order
+    let [mut msg0, mut msg1, mut msg2, mut msg3] = [0, 1, 2, 3].map(|i| {
+        let p = block.as_ptr().add(16 * i);
+        vreinterpretq_u32_u8(vrev32q_u8(vld1q_u8(p)))
+    });
+
+    vst1q_u32(wp, msg0);
+    vst1q_u32(wp.add(4), msg1);
+    vst1q_u32(wp.add(8), msg2);
+    vst1q_u32(wp.add(12), msg3);
+
+    tmp0 = vaddq_u32(msg0, k0);
+    tmp1 = vaddq_u32(msg1, k0);
+
+    // Rounds 0-3
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1cq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg2, k0);
+    msg0 = vsha1su0q_u32(msg0, msg1, msg2);
+
+    // Rounds 4-7
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1cq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg3, k0);
+    msg0 = vsha1su1q_u32(msg0, msg3);
+    vst1q_u32(wp.add(16), msg0);
+    msg1 = vsha1su0q_u32(msg1, msg2, msg3);
+
+    // Rounds 8-11
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1cq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg0, k0);
+    msg1 = vsha1su1q_u32(msg1, msg0);
+    vst1q_u32(wp.add(20), msg1);
+    msg2 = vsha1su0q_u32(msg2, msg3, msg0);
+
+    // Rounds 12-15
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1cq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg1, k1);
+    msg2 = vsha1su1q_u32(msg2, msg1);
+    vst1q_u32(wp.add(24), msg2);
+    msg3 = vsha1su0q_u32(msg3, msg0, msg1);
+
+    // Rounds 16-19
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1cq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg2, k1);
+    msg3 = vsha1su1q_u32(msg3, msg2);
+    vst1q_u32(wp.add(28), msg3);
+    msg0 = vsha1su0q_u32(msg0, msg1, msg2);
+
+    // Rounds 20-23
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg3, k1);
+    msg0 = vsha1su1q_u32(msg0, msg3);
+    vst1q_u32(wp.add(32), msg0);
+    msg1 = vsha1su0q_u32(msg1, msg2, msg3);
+
+    // Rounds 24-27
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg0, k1);
+    msg1 = vsha1su1q_u32(msg1, msg0);
+    vst1q_u32(wp.add(36), msg1);
+    msg2 = vsha1su0q_u32(msg2, msg3, msg0);
+
+    // Rounds 28-31
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg1, k1);
+    msg2 = vsha1su1q_u32(msg2, msg1);
+    vst1q_u32(wp.add(40), msg2);
+    msg3 = vsha1su0q_u32(msg3, msg0, msg1);
+
+    // Rounds 32-35
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg2, k2);
+    msg3 = vsha1su1q_u32(msg3, msg2);
+    vst1q_u32(wp.add(44), msg3);
+    msg0 = vsha1su0q_u32(msg0, msg1, msg2);
+
+    // Rounds 36-39
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg3, k2);
+    msg0 = vsha1su1q_u32(msg0, msg3);
+    vst1q_u32(wp.add(48), msg0);
+    msg1 = vsha1su0q_u32(msg1, msg2, msg3);
+
+    // Rounds 40-43
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1mq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg0, k2);
+    msg1 = vsha1su1q_u32(msg1, msg0);
+    vst1q_u32(wp.add(52), msg1);
+    msg2 = vsha1su0q_u32(msg2, msg3, msg0);
+
+    // Rounds 44-47
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1mq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg1, k2);
+    msg2 = vsha1su1q_u32(msg2, msg1);
+    vst1q_u32(wp.add(56), msg2);
+    msg3 = vsha1su0q_u32(msg3, msg0, msg1);
+
+    // Rounds 48-51
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1mq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg2, k2);
+    msg3 = vsha1su1q_u32(msg3, msg2);
+    vst1q_u32(wp.add(60), msg3);
+    msg0 = vsha1su0q_u32(msg0, msg1, msg2);
+
+    // Rounds 52-55
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1mq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg3, k3);
+    msg0 = vsha1su1q_u32(msg0, msg3);
+    vst1q_u32(wp.add(64), msg0);
+    msg1 = vsha1su0q_u32(msg1, msg2, msg3);
+
+    // Rounds 56-59
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1mq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg0, k3);
+    msg1 = vsha1su1q_u32(msg1, msg0);
+    vst1q_u32(wp.add(68), msg1);
+    msg2 = vsha1su0q_u32(msg2, msg3, msg0);
+
+    // Rounds 60-63
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg1, k3);
+    msg2 = vsha1su1q_u32(msg2, msg1);
+    vst1q_u32(wp.add(72), msg2);
+    msg3 = vsha1su0q_u32(msg3, msg0, msg1);
+
+    // Rounds 64-67
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e0, tmp0);
+    tmp0 = vaddq_u32(msg2, k3);
+    msg3 = vsha1su1q_u32(msg3, msg2);
+    vst1q_u32(wp.add(76), msg3);
+
+    // Rounds 68-71
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e1, tmp1);
+    tmp1 = vaddq_u32(msg3, k3);
+
+    // Rounds 72-75
+    e1 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e0, tmp0);
+
+    // Rounds 76-79
+    e0 = vsha1h_u32(vgetq_lane_u32(abcd, 0));
+    abcd = vsha1pq_u32(abcd, e1, tmp1);
+
+    abcd = vaddq_u32(abcd_cpy, abcd);
+    e0 = e0.wrapping_add(e0_cpy);
+
+    vst1q_u32(state.as_mut_ptr(), abcd);
+    state[4] = e0;
+}
